Cleanup
Accept linter suggestions
Remove "find card by suit and value" as this can be done with filter
Remove "remove card from hand by suit and value" as the card should be known before removing it.
Allow the game engine to skip a player's turn if there are too many cards in play
* The only case where failure to deal is an explicit error is if setup fails to deal enough cards initially.
Predict winner as a card is played
Reorganize turn handling, so that it's more obvious that advancePlayer() will always be called.
Use listOfNotNull(loser) for cleaner handling of null loser